+++ packages/cli/src/util/deploy/process-deployment.ts
@@ -289,15 +289,13 @@ export default async function processDeployment({
         return event.payload;
       }
 
-      // If `checksState` is present, we can only continue to "Completing" if the checks finished,
-      // otherwise we might show "Completing" before "Running Checks".
-      if (
-        event.type === 'ready' &&
-        (event.payload.checksState
-          ? event.payload.checksState === 'completed'
-          : true) &&
-        !withFullLogs
-      ) {
+      if (event.type === 'ready' && !withFullLogs) {
+        const v1ChecksPending =
+          event.payload.checksState &&
+          event.payload.checksState !== 'completed';
+        const v2ChecksPending =
+          event.payload.checks?.['deployment-alias']?.state === 'pending';
+
         stopSpinner();
         process.stderr.write(eraseLines(2));
         const isProdDeployment = event.payload.target === 'production';
@@ -310,14 +308,26 @@ export default async function processDeployment({
             emoji('success')
           ) + `\n`
         );
-        output.spinner('Completing...', 0);
+
+        if (v1ChecksPending || v2ChecksPending) {
+          output.spinner('Running Checks...', 0);
+        } else {
+          output.spinner('Completing...', 0);
+        }
       }
 
+      // v1 checks running
       if (event.type === 'checks-running' && !withFullLogs) {
         output.spinner('Running Checks...', 0);
       }
 
+      // v1 checks failed
       if (event.type === 'checks-conclusion-failed') {
+        stopSpinner();
+        return event.payload;
+      }
+
+      if (event.type === 'checks-v2-failed') {
         stopSpinner();
         return event.payload;
       }